==> Exam/Code/2. Data_rens.ipynb <==
#Import packages
import pandas as pd
import re
import numpy as np
from sklearn.feature_extraction.text import TfidfVectorizer
from sklearn.linear_model import LogisticRegression
from pathlib import Path

import nltk
nltk.download('stopwords')
from nltk.corpus import stopwords

stop = stopwords.words('english')

nltk.download('omw-1.4') #Download OpenMultilingualWordnet
nltk.download('wordnet')
wnl = nltk.WordNetLemmatizer()


fp_data = Path.cwd() / "data"
Path.mkdir(fp_data, exist_ok=True) 

# Import data
df = pd.read_csv(fp_data / 'bookssave100.csv')
df=df.drop('Unnamed: 0', axis=1)
df = df.dropna(axis='rows').reset_index(drop=True)

#droping duplicates 
df=df.drop_duplicates().reset_index(drop=True)

#Defines function that takes a text data column and processes it.
def preprocess(column):
    column = column.lower() #To lower case
    column = re.sub(r'<[^>]*>', ' ', column) #Remove HTML
    column = re.sub(r'[^\w\s]','', column) #Remove non-alphanumeric characters

    q = column.split() #Split sentences at whitespace
    w = [e for e in q if e not in stop] #Removes "stop-word" e.g. "The" and "is"
    u = [wnl.lemmatize(e) for e in w] #Finds the grammatically correct form of the word (the lemma).
    return u

# Renames the old columns to the new processed columns
df_be = df['Description'].apply(preprocess) 
df_ti = df['Title'].apply(preprocess)

df_fin = df.copy()
df_fin['proc_de'] = df_be
df_fin['proc_ti'] = df_ti

#Dropping books without Author specified 
df_fin = df_fin.drop(df_fin[df_fin['Author'].str.contains('Academy', case=False, na=False)].index)
df_fin = df_fin.drop(df_fin[df_fin['Author'].str.contains('Science Association', case=False, na=False)].index)

#Splits the string containing the names of the authors, into a list of strings, each string containing the name of one author.
df_fin['names_column'] = df_fin['Author'].str.split(',')

#Drop if there is more than 4 Authors
df_fin = df_fin[df_fin['names_column'].apply(len) <= 4]
df_fin = df_fin.reset_index(drop=True)
#Splits the name column into three columns, each containing one name.
df_fin[['Author1', 'Author2', 'Author3', 'Author4']] = pd.DataFrame(df_fin['names_column'].tolist())

# Deletes all the names that are not authors, i.e. the names with a parenthesis afterwards.
import gender_guesser.detector as gender
d = gender.Detector()

def process_name(name):
    if pd.notna(name):
        if '(' in name or ')' in name:
            return None
        
        stripped_name = name.strip()
        parts = stripped_name.split(' ', 1)
        first_name = parts[0]
        
        gender = d.get_gender(first_name)  # Assuming d is defined somewhere

        return gender
    else:
        return None

# Applies the function to the name columns
df_fin['Author1'] = df_fin['Author1'].apply(lambda x: process_name(x))
df_fin['Author2'] = df_fin['Author2'].apply(lambda x: process_name(x))
df_fin['Author3'] = df_fin['Author3'].apply(lambda x: process_name(x))
df_fin['Author4'] = df_fin['Author4'].apply(lambda x: process_name(x))

df_fin = df_fin.drop(df_fin[df_fin['Author1'].str.contains('unknown', case=False, na=False)].index)
df_fin = df_fin.drop(df_fin[df_fin['Author2'].str.contains('unknown', case=False, na=False)].index)
df_fin = df_fin.drop(df_fin[df_fin['Author3'].str.contains('unknown', case=False, na=False)].index)
df_fin = df_fin.drop(df_fin[df_fin['Author4'].str.contains('unknown', case=False, na=False)].index)

# Replaces the words of the genders with numbers.
gender_mapping = {'male': 1, 'female': 2, 'unknown': 3, 'mostly_male': 1, 'mostly_female': 2}
df_fin['Author1'] = df_fin['Author1'].map(gender_mapping)     
df_fin['Author2'] = df_fin['Author2'].map(gender_mapping)     
df_fin['Author3'] = df_fin['Author3'].map(gender_mapping)
df_fin['Author4'] = df_fin['Author4'].map(gender_mapping)
# Replaces empty name cells with the value 0
df_fin['Author1'].fillna(value=0, inplace=True)
df_fin['Author2'].fillna(value=0, inplace=True)
df_fin['Author3'].fillna(value=0, inplace=True)
df_fin['Author4'].fillna(value=0, inplace=True)

# Drop if published year is outlier
df_fin = df_fin.drop(df_fin[df_fin['Published_year'] < 1001].index)
df_fin = df_fin.drop(df_fin[df_fin['Published_year'] > 2023].index)

mask = (df_fin['Author1'] == 0) & (df_fin['Author2'] == 0) & (df_fin['Author3'] == 0) & (df_fin['Author4'] == 0)
df_fin = df_fin.drop(df_fin[mask].index)
df_fin

df_fin = df_fin.reset_index(drop=True)
df_fin

df_fin = df_fin.drop(columns=['Title', 'Author', 'Description', 'names_column'])

df_fin.to_csv('data_processed.csv', index=False) #Saves the processed data as a csv-file.
